refactor: files refactoring & moving
