+++ src/components/pages/ProjectsPage/ProjectsCreateModal/ProjectsCreateModal.tsx
@@ -3,18 +3,19 @@ import React, { FC } from 'react';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { format } from 'date-fns';
-import { Spinner } from '../../../components/Spinner';
-import { InputText } from '../../../components/UI/InputText';
-import { IUserAllResult } from '../../../interfaces/IUser.interface';
-import { ModalWindow } from '../../../components/UI/ModalWindow';
-import { TFormSubmit } from '../../../types/formTypes';
-import { TError } from '../../../types/errorTypes';
-import { CREATE_PROJECT } from '../../../graphql/mutations/createProject/createProject';
-import { FieldNameProjectsForm } from '../../../constants/FieldNameProjectsForm';
-import { projectsSchema } from '../../../utils/validationSchema';
-import { CreateProjectResult } from '../../../graphql/mutations/createProject/createProject.types';
-import { updateCacheAfterCreatingProject } from '../../../graphql/mutations/createProject/createProject.cache';
-import { DatePickerInput } from '../../../components/UI/DatePicker';
+
+import { Spinner } from '../../../Spinner';
+import { InputText } from '../../../UI/InputText';
+import { IUserAllResult } from '../../../../graphql/types/results/user';
+import { ModalWindow } from '../../../UI/ModalWindow';
+import { DatePickerInput } from '../../../UI/DatePicker';
+import { updateCacheAfterCreatingProject } from '../../../../graphql/cache/createProject';
+import { CREATE_PROJECT } from '../../../../graphql/mutations/createProject';
+import { projectsSchema } from '../../../../utils/validationSchema';
+import { CreateProjectResult } from '../../../../graphql/types/results/projects';
+import { TError } from '../../../../types/errorTypes';
+import { FieldNameProjectsForm } from '../../../../constants/FieldNameProjectsForm';
+import { TFormSubmit } from '../../../../types/formTypes';
 import * as Styled from './../../EmployeesPage/EmployeesModal/EmployeesModal.styles';
 import { IProjectsFormInput, IProjectsModalProps } from './ProjectsCreateModal.interface';
