+++ src/components/pages/ProjectsPage/ProjectsPage.tsx
@@ -2,13 +2,13 @@ import { useQuery } from '@apollo/client';
 import React from 'react';
 import { Container, Grid } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
-import { createTable } from '../../components/Table/template/templateTable';
 
-import { Spinner } from '../../components/Spinner';
-import { RoutePath } from '../../constants/routeVariables';
-import { GET_ALL_PROJECTS } from '../../graphql/queries/projects';
-import { useUser } from '../../hooks/useUser';
-import { UserRoles } from '../../constants/userRoles';
+import { RoutePath } from '../../../constants/routeVariables';
+import { Spinner } from '../../Spinner';
+import { createTable } from '../../Table/template';
+import { GET_ALL_PROJECTS } from '../../../graphql/queries/projects';
+import { useUser } from '../../../hooks/useUser';
+import { UserRoles } from '../../../constants/userRoles';
 import { ProjectsTableHeader } from './TableData/ProjectsTableHeader';
 import { getProjects } from './TableData/ProjectsTableRows';
 import { ProjectsAdditionalButtons } from './ProjectsAdditionalButtons/ProjectsAdditionalButtons';
