+++ src/graphql/mutations/deleteProject.ts
