--- src/graphql/mutations/deleteProject/deleteProject.types.ts
@@ -1,13 +1,0 @@
-export type DeleteProjectInput = {
-  id: string;
-};
-
-export type DeleteProjectResult = {
-  deleteProject: {
-    affected: number;
-  };
-};
-
-export type DeleteResult = {
-  affected: number;
-};
