+++ src/graphql/types/results/projects.ts
@@ -1,0 +1,24 @@
+import { IProject } from '../../../interfaces/IProject.interface';
+import { ProjectInput } from '../inputs/projects';
+
+export type DeleteProjectResult = {
+  deleteProject: {
+    affected: number;
+  };
+};
+
+export type DeleteResult = {
+  affected: number;
+};
+
+export interface IProjectsResult {
+  projects: IProject[];
+}
+
+export interface CreateProjectResult {
+  createProject: {
+    id: string;
+    project: ProjectInput;
+    affected: number;
+  };
+}
